fix: fixed translations
