+++ src/components/layout/header/controls-desktop.astro
@@ -1,7 +1,7 @@
 ---
 import { languages } from '@/lib/constants'
 import { ChevronDown, Sun, Moon } from 'lucide-react'
-import { t } from '@/lib/astro-i18n'
+import { t } from '@/lib/i18n-server'
 ---
 
 <div class="flex items-center space-x-6">
