+++ src/components/layout/header/controls-mobile.astro
@@ -1,7 +1,7 @@
 ---
 import { languages } from '@/lib/constants'
 import { Sun, Moon } from 'lucide-react'
-import { t } from '@/lib/astro-i18n'
+import { t } from '@/lib/i18n-server'
 ---
 
 <div class="border-t border-gray-200 pt-4 dark:border-gray-700">
