+++ src/components/layout/header/mobile-button.astro
@@ -1,6 +1,6 @@
 ---
 import { Menu, X } from 'lucide-react'
-import { t } from '@/lib/astro-i18n'
+import { t } from '@/lib/i18n-server'
 
 interface Props {
 	action: 'open' | 'close'
