+++ src/components/layout/header/navigation.astro
@@ -1,7 +1,7 @@
 ---
 import { cn } from '@/lib/utils'
 import { ChevronRight } from 'lucide-react'
-import { t } from '@/lib/astro-i18n'
+import { t } from '@/lib/i18n-server'
 
 interface Props {
 	links: { href: string; labelKey: string }[]
