+++ src/components/legal/legal-content.astro
@@ -1,5 +1,5 @@
 ---
-import { t } from '@/lib/astro-i18n'
+import { t } from '@/lib/i18n-server'
 import GradientCircle from '@/components/common/gradient-circle.astro'
 
 export interface Props {
