+++ src/components/marketing/hero-content.astro
@@ -2,7 +2,7 @@
 import { Button } from '@/components/ui/button'
 import { cn } from '@/lib/utils'
 import { StartProjectModal } from './start-project-modal'
-import { t } from '@/lib/astro-i18n'
+import { t } from '@/lib/i18n-server'
 ---
 
 <!-- Main Content -->
