+++ src/components/marketing/scroll-indicator.astro
@@ -1,6 +1,6 @@
 ---
 import { cn } from '@/lib/utils'
-import { t } from '@/lib/astro-i18n'
+import { t } from '@/lib/i18n-server'
 ---
 
 <!-- Scroll Indicator - Mobile Only -->
