+++ src/components/marketing/start-project-modal.tsx
@@ -1,5 +1,5 @@
 import { cn } from '@/lib/utils'
-import { useI18n } from '@/lib/i18n'
+import { useI18n } from '@/lib/i18n-client'
 import {
 	Dialog,
 	DialogContent,
