+++ src/pages/privacy.astro
@@ -4,7 +4,7 @@ import Header from '@/components/layout/header.astro'
 import Footer from '@/components/layout/footer.astro'
 import LegalContent from '@/components/legal/legal-content.astro'
 import Section from '@/components/legal/section.astro'
-import { t } from '@/lib/astro-i18n'
+import { t } from '@/lib/i18n-server'
 
 // i18n est déjà initialisé par le middleware
 const pageTitle = t('privacy.title')
